trunk/ (mint@mint) BUG FIX: starting to fix compiler offsets on 64-bit Linux machines. Fixed compoud offset assignment for generic types, which assumed that sizeof(int) == sizeof(ssize_t). Several list tests pass again, but two still fail.
git-svn-id: svn+ssh://saulius-grazulis.lt/home/saulius/svn-repositories/compilers/trunk@2969 f7c8cbe2-f9e6-0310-9050-d7f47bf15973

+++ sc/dnode.c
@@ -709,7 +709,7 @@ DNODE *dnode_list_last( DNODE *dnode )
     }
 }
 
-void dnode_assign_offset( DNODE *dnode, int *offset )
+void dnode_assign_offset( DNODE *dnode, ssize_t *offset )
 {
     int delta;
     ssize_t stackcells;
@@ -723,7 +723,7 @@ void dnode_assign_offset( DNODE *dnode, int *offset )
     dnode->offset = *offset - delta;
 }
 
-void dnode_list_assign_offsets( DNODE *dnode_list, int *offset )
+void dnode_list_assign_offsets( DNODE *dnode_list, ssize_t *offset )
 {
     DNODE *dnode;
 
@@ -733,7 +733,7 @@ void dnode_list_assign_offsets( DNODE *dnode_list, int *offset )
     }
 }
 
-DNODE *dnode_set_offset( DNODE *dnode, int offset )
+DNODE *dnode_set_offset( DNODE *dnode, ssize_t offset )
 {
     assert( dnode );
 #if 0
@@ -745,7 +745,7 @@ DNODE *dnode_set_offset( DNODE *dnode, int offset )
     return dnode;
 }
 
-DNODE *dnode_update_offset( DNODE *dnode, int offset )
+DNODE *dnode_update_offset( DNODE *dnode, ssize_t offset )
 {
     assert( dnode );
     dnode->offset = offset;
